initial

+++ src/main/java/za/ac/cput/repository/IOrderItemRepository.java
@@ -3,5 +3,5 @@ package za.ac.cput.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import za.ac.cput.domain.OrderItem;
 
-public interface IOrderItemRepository extends JpaRepository<OrderItem, Integer> {
+public interface IOrderItemRepository extends JpaRepository<OrderItem, Long> {
 }
